// src/app/models/site.ts
export interface Site {
    id: number | null;
    datcrea: string;
    datmodi: string;
    paycode: string;
    quicrea: string;
    quimodi: string;
    reptiecode: string;
    sercode: string;
    sitadr1: string;
    sitadr2: string;
    sitcode: string;
    sitcp: string;
    sitcptembdest: string;
    sitcptembexp: string;
    sitembsercode: string;
    sitembsitcode: string;
    sitembtiepay: string;
    sitfax: string;
    sitgtf: string;
    sitlibl: string;
    sitmail: string;
    sitretour: boolean;
    sitserarr: string;
    sitserret: string;
    sitsiretedi: string;
    sittel: string;
    sitville: string;
    soccode: string;
    vilident: string;
    latitude: number | null; // Added
    longitude: number | null; // Added
  }